+++ Problem327/solve.hpp
@@ -4,4 +4,4 @@
 /**
  * Returns the number of cards pulled from the dispensor, in order to carry K cards through R doors, carrying at most C cards at a time.
  */
-int solve(int C, int R, int K = 1);
+unsigned long solve(unsigned long C, unsigned long R, unsigned long K = 1);
